# Avoid calling merge_or_append to speed up 