--- a/lib/new_relic/agent/range_extensions.rb
+++ b/lib/new_relic/agent/range_extensions.rb
@@ -12,26 +12,6 @@ module NewRelic
         r1.include?(r2.begin) || r2.include?(r1.begin)
       end
 
-      def merge(r1, r2)
-        return unless intersects?(r1, r2)
-        range_min = r1.begin < r2.begin ? r1.begin : r2.begin
-        range_max = r1.end > r2.end ? r1.end : r2.end
-        range_min..range_max
-      end
-
-      # Takes an array of ranges and a range which it will
-      # merge into an existing range if they intersect, otherwise
-      # it will append this range to the end the array.
-      def merge_or_append(range, ranges)
-        ranges.each_with_index do |r, i|
-          if merged = merge(r, range)
-            ranges[i] = merged
-            return ranges
-          end
-        end
-        ranges.push(range)
-      end
-
       # Computes the amount of overlap between range and an array of ranges.
       # For efficiency, it assumes that range intersects with each of the
       # ranges in the ranges array.