--- a/test/new_relic/agent/range_extensions_test.rb
+++ b/test/new_relic/agent/range_extensions_test.rb
@@ -17,26 +17,6 @@ module NewRelic
         refute RangeExtensions.intersects?((1...3), (3..5))
       end
 
-      def test_merge_merges_intersecting_ranges
-        merged = RangeExtensions.merge((1..5), (3..8))
-        assert_equal (1..8), merged
-      end
-
-      def test_merge_nil_for_disjoint_ranges
-        merged = RangeExtensions.merge((1...3), (3..5))
-        assert_nil merged
-      end
-
-      def test_merge_or_append_merges_intersecting_ranges
-        result = RangeExtensions.merge_or_append((3..8), [(1..5), (9..13)])
-        assert_equal [(1..8), (9..13)], result
-      end
-
-      def test_merge_or_append_appends_disjoint_ranges
-        result = RangeExtensions.merge_or_append((6..8), [(1..5), (9..13)])
-        assert_equal [(1..5), (9..13), (6..8)], result
-      end
-
       def test_compute_overlap
         result = RangeExtensions.compute_overlap(0..10, [-4..2, 6..7, 9..15])
         assert_equal 4, result