_Before contributing, please read our [contributing guidelines](https://github.com/newrelic/newrelic-ruby-agent/blob/main/CONTRIBUTING.md) and [code of conduct](https://github.com/newrelic/.github/blob/master/CODE_OF_CONDUCT.md)._

I've read them 👍 

# Overview

newrelic_rpm gets slow when `NewRelic::Agent::RangeExtensions.merge_or_append` is called so many times.

https://github.com/newrelic/newrelic-ruby-agent/issues/1264

`newrelic_rpm@8.9.0` slows the response, 446ms -> 33670ms. (See https://github.com/hkdnet/graphql-ruby-nr/tree/ccefc10dcae5a1ed90eaba4cc6f8ea3bcb6e38a4#result )
But the current patch slows the same response, 403ms -> 626ms. (See https://github.com/hkdnet/graphql-ruby-nr/tree/perf-test-for-lazy-merge#result )

Though the result is not stable (it depends on my machine's status), it shows a significant improvement.

Submitter Checklist:
- [x] Include a link to the related GitHub issue, if applicable
- [ ] Include a security review link, if applicable
    - I think we don't need it.

# Testing
The agent includes a suite of unit and functional tests which should be used to
verify your changes don't break existing functionality. These tests will run with 
Github Actions when a pull request is made. More details on running the tests locally can be found 
[here for our unit tests](https://github.com/newrelic/newrelic-ruby-agent/blob/main/test/README.md), 
and [here for our functional tests](https://github.com/newrelic/newrelic-ruby-agent/blob/main/test/multiverse/README.md).
For most contributions it is strongly recommended to add additional tests which
exercise your changes. 

# Reviewer Checklist
- [ ] Perform code review
- [ ] Add performance label
- [ ] Perform appropriate level of performance testing
- [ ] Confirm all checks passed
- [ ] Add version label prior to acceptance