@fallwith 2022-07-20
Hi @hkdnet! Thank you so much for your efforts to get the `graphql` and `newrelic_rpm` gems to better harmonize. We love GraphQL and we want to provide GraphQL adopters with useful information without incurring performance overhead.

The `newrelic_rpm` gem keeps track of several Range objects in order to help with multiple concurrently running traces. In particular, the [test_total_time_metrics_async_sync_children_non_web](https://github.com/newrelic/newrelic-ruby-agent/blob/dev/test/new_relic/agent/transaction/tracing_test.rb#L810) has comments with an illustration to explain the desired functionality.

It seems like the `merge_or_append` method is the main cause of performance problems for `graphql`. If that method goes away, it seems that some core `newrelic_rpm` trace functionality stops working and some tests fail.

Thank you very much for moving your PR to draft status and looking at the tests. We will also continue looking at this problem.

One current hypothesis that we have is that `graphql`'s use of [trace_execution_scoped](https://github.com/rmosolgo/graphql-ruby/blob/master/lib/graphql/tracing/new_relic_tracing.rb#L33) is possibly getting called too frequently, especially when Dataloader is involved. If that method is called too frequently, it could cause there to be too many Range objects concurrently being tracked by `newrelic_rpm` and performance problems take place.

We definitely want to see what can be done to improve the performance of handling Range objects in `newrelic_rpm`, and if possible we would also like some help from the `graphql` community in evaluating ways to possibly improve the use of `trace_execution_scoped` when Dataloader is engaged.