@hkdnet 2022-07-30
https://github.com/newrelic/newrelic-ruby-agent/pull/1274 covers this 👋